backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
twilio==9.9.0
qdrant-client==1.16.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
requests==2.32.5
Flask==3.1.2
python-magic==0.4.27        # ✅ NEW
python-magic-bin==0.4.14    # ✅ NEW (Windows)
Pillow==10.2.0         # ✅ NEW